* commit 'fa49d986f61109e370a0ecaa7645609da472354b': (50 commits)
  REPORT-26147 @harrison：新的release的jar和旧版本的性能插件，性能插件需要自动禁用 改 apilevel 满足不了全部禁用的需求。 改回原来的名字，不然旧版本的性能插件没法正常用。
  REPORT-26091 && REPORT-26181
  REPORT-26160 && REPORT-26204 && && REPORT-25532
  MOBILE-23781 单选按钮组的弹出配置框,有时候点击Tab不会切换
  MOBILE-24787 设计器上第一次打开已有弹窗设置模板，发现弹窗设置界面为空
  REPORT-24314 10.0.4启动画面 非中文版提供
  REPORT-25645 mac设计器取色器异常
  REPORT-25447 10.0.4设计器表单设计奇怪的bug
  MOBILE-24213 单选按钮 选项值前面会出现虚线 && 默认值
  判断条件有误
  REPORT-25730 最新的release设计器启动后一直报模板不存在 还需要判断是否存在
  REPORT-25730 最新的release设计器启动后一直报模板不存在 这里执行的时候需要判空。
  无jira任务, 冲突
  调整
  无JIRA任务，1.8编译单元测试代码失败
  fix
  REPORT-25417 智能添加单元格点击不准确
  1.8编译单元测试代码失败
  REPORT-24604 表单拖动图表与画布块交叉设计器卡死
  REPORT-24344 Spider数据集插件需要约束下BI 相关jar的 jar time 开放校验接口。
  ...

+++ designer-base/src/test/java/com/fr/design/mainframe/template/info/TemplateInfoTest.java
@@ -54,7 +54,7 @@ public class TemplateInfoTest {
         TemplateInfo templateInfo = TemplateInfo.newInstance(templateID);
         assertEquals(templateID, templateInfo.getTemplateID());
         assertEquals(StringUtils.EMPTY, Reflect.on(templateInfo).field("originID").get());
-        assertEquals(0, Reflect.on(templateInfo).field("idleDayCount").get());
+        assertEquals(0, (int) Reflect.on(templateInfo).field("idleDayCount").get());
         assertEquals("{}", templateInfo.getProcessMapJsonString());
 
         Map<String, Object> consumingMap = Reflect.on(templateInfo).field("consumingMap").get();
@@ -77,7 +77,7 @@ public class TemplateInfoTest {
         TemplateInfo templateInfo = TemplateInfo.newInstance(templateID, originID, originTime);
         assertEquals(templateID, templateInfo.getTemplateID());
         assertEquals(originID, Reflect.on(templateInfo).field("originID").get());
-        assertEquals(0, Reflect.on(templateInfo).field("idleDayCount").get());
+        assertEquals(0, (int) Reflect.on(templateInfo).field("idleDayCount").get());
         assertEquals("{}", templateInfo.getProcessMapJsonString());
 
         Map<String, Object> consumingMap = Reflect.on(templateInfo).field("consumingMap").get();
